requirements.txt
Django
django-filter
djangorestframework
model-bakery
psycopg2-binary
pytest-cov
pytest-django
PyYAML
